Replaced headlessui popover with shadcn popover in UpdatableApp; Uninstall headlessui (#10900)
