+++ package.json
@@ -52,7 +52,6 @@
   "dependencies": {
     "@excalidraw/excalidraw": "^0.17.6",
     "@fontsource/figtree": "^5.1.1",
-    "@headlessui/react": "^2.2.0",
     "@hookform/resolvers": "^4.0.0",
     "@originjs/vite-plugin-federation": "^1.3.7",
     "@radix-ui/react-alert-dialog": "^1.1.2",
