+++ public/locale/en.json
@@ -270,6 +270,7 @@
   "VENTILATOR_MODE__VC_SIMV": "Volume Controlled SIMV (VC-SIMV)",
   "VENTILATOR_MODE__VC_SIMV_short": "VC-SIMV",
   "View_Facility": "View Facility",
+  "a_new_version_of_care_is_available": "A new version of CARE is available",
   "aadhaar_number": "Aadhaar Number",
   "aadhaar_number_will_not_be_stored": "Aadhaar number will not be stored by CARE",
   "aadhaar_otp_send_error": "Failed to send OTP. Please try again later.",
@@ -394,6 +395,7 @@
   "any_additional_information": "Any Additional Information",
   "any_other_comments": "Any other comments",
   "app_settings": "App Settings",
+  "application_successfully_updated": "Application successfully updated!",
   "apply": "Apply",
   "appointment_booking_success": "Your appointment has been successfully booked!",
   "appointment_cancelled": "Appointment has been cancelled!",
@@ -1732,6 +1734,7 @@
   "notification_permission_denied": "Notification permission denied",
   "notification_permission_granted": "Notification permission granted",
   "notify": "Notify",
+  "now_using_the_latest_version_of_care": "Now using the latest version of CARE",
   "number_min_error": "Must be greater than {{min}}",
   "number_of_aged_dependents": "Number of Aged Dependents (Above 60)",
   "number_of_beds": "Number of beds",
@@ -2610,8 +2613,10 @@
   "update_volunteer": "Reassign Volunteer",
   "updated": "Updated",
   "updated_on": "Updated On",
+  "updated_successfully": "Updated successfully",
   "updates": "Updates",
   "updating": "Updating...",
+  "updating_application": "Updating application...",
   "updating_facility": "Updating Facility...",
   "upload": "Upload",
   "upload_an_image": "Upload an image",
